feat : possibility to change the newline split for different file format
debug : couldn't generate a new file from basic json data

Feat description :
When using a file format that does newline as '\r\n' and not '\n', the "splitRecords" function wasn't able to split correctly the records. SO I added a possibility to set externally the newline split string.

Debug description :
the "Options" property in the record object wasn't included in the parse to RecordObject, which then wasn't taken into account upon generating new db.local file.
Also, the generated custom Object wasn't the same as the documentation, because of lowercase keys and the "Option" property